image upscaling nvidia/requirements.txt
# PyTorch with CUDA 11.8 support (for NVIDIA GPUs)
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.0.0
tqdm==4.66.1
tensorboard==2.14.0
opencv-python==4.8.1.78
scikit-image==0.21.0
matplotlib==3.8.0
scipy==1.11.3
